requirements.txt
# Dependencies for Storage Topology Tool

# YAML configuration file support
PyYAML>=6.0
